+++ storage/db-core/src/main/java/com/barlow/storage/db/core/LegislationAccountJpaEntity.java
@@ -1,6 +1,6 @@
 package com.barlow.storage.db.core;
 
-import com.barlow.core.domain.account.LegislationAccount;
+import com.barlow.core.domain.legislationaccount.LegislationAccount;
 import com.barlow.core.domain.home.MyLegislationAccount;
 
 import jakarta.persistence.Column;
