+++ storage/db-core/src/main/java/com/barlow/storage/db/core/NotificationConfigJpaEntity.java
@@ -1,7 +1,7 @@
 package com.barlow.storage.db.core;
 
 import com.barlow.core.domain.User;
-import com.barlow.core.domain.account.notificationsetting.NotificationSetting;
+import com.barlow.core.domain.legislationaccount.notificationsetting.NotificationSetting;
 import com.barlow.core.domain.notification.NotificationConfig;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
